requirements.txt
streamlit==1.38
pandas
azure-cosmos
azure-storage-blob